Consolidate modules
