--- src/rhsmlib/dbus/common/exceptions.py
@@ -1,3 +1,0 @@
-
-class Error(Exception):
-    pass
