+++ src/rhsmlib/dbus/services/examples/rhsm-examples-service
@@ -1,18 +1,34 @@
-#!/usr/bin/python
-
+#!/usr/bin/env python
+import dbus
 import sys
-
-sys.path.append("/usr/share/rhsm")
+import optparse
 
 from subscription_manager import logutil
 logutil.init_logger()
 
-from rhsmlib.dbus.services.examples import server
+from rhsmlib import import_class
+from rhsmlib.dbus.services import base_server
+from rhsmlib.dbus.services.examples import examples
 
 
 def main(args):
-    server.run()
+    bus_class = dbus.SystemBus
+    bus_name = examples.DBUS_NAME
+
+    if len(args) > 0:
+        bus_class = import_class(args[0])
+
+    if len(args) > 1:
+        bus_name = args[1]
+
+    service_classes = [examples.Examples]
+
+    base_server.run_services(bus_class=bus_class,
+        bus_name=bus_name,
+        service_classes=service_classes)
     return 0
 
 if __name__ == "__main__":
-    sys.exit(main(sys.argv))
+    parser = optparse.OptionParser()
+    (options, args) = parser.parse_args()
+    sys.exit(main(args))
