+++ src/rhsmlib/facts/virt.py
@@ -21,7 +21,7 @@ import logging
 from rhsmlib.facts import collector
 
 # For python2.6 that doesn't have subprocess.check_output
-from rhsmlib.compat.subprocess_check_output import check_output as compat_check_output
+from rhsmlib.compat import check_output as compat_check_output
 
 log = logging.getLogger(__name__)
